Feature/local cart (#17)
* added a local cart

* changes on subtract local cart items

* added a bulk add cart item action,service

* changes on bulk add cart drequest data

+++ Fastkart-Front-SSR/src/app/components/shop/cart/cart.component.ts
@@ -4,7 +4,7 @@ import { Observable } from 'rxjs';
 import { Breadcrumb } from '../../../shared/interface/breadcrumb';
 import { Cart, CartAddOrUpdate } from '../../../shared/interface/cart.interface';
 import { CartState } from '../../../shared/state/cart.state';
-import { UpdateCart, DeleteCart, AddToCart } from '../../../shared/action/cart.action';
+import { UpdateCart, DeleteCart, AddToCart, SubractFromCartLocalStorage, AddToCartLocalStorage } from '../../../shared/action/cart.action';
 import { AddOrRemoveWishlist } from '../../../shared/action/wishlist.action';
 import { convertStringToNumber, getStringDataFromLocalStorage } from 'src/app/utilities/helper';
 import { environment } from 'src/environments/environment';
@@ -16,7 +16,7 @@ import { environment } from 'src/environments/environment';
 })
 export class CartComponent {
   cartItems: Array<any> = [];
-  apiBaseUrl:string=environment.apiBaseUrl;
+  apiBaseUrl: string = environment.apiBaseUrl;
   @Select(CartState.cartItems) cartItem$: Observable<Cart[]>;
   @Select(CartState.cartTotal) cartTotal$: Observable<number>;
 
@@ -50,7 +50,16 @@ export class CartComponent {
   }
 
   delete(itemId: number) {
-    this.store.dispatch(new DeleteCart({ id: itemId }));
+    const user_id = getStringDataFromLocalStorage("user_id");
+
+    if (user_id) {
+      this.store.dispatch(new DeleteCart({ id: itemId }));
+    } else {
+      let data = this.cartItems.find((e: any) => e.id === itemId);
+      data.qty = 0;
+
+      this.store.dispatch(new SubractFromCartLocalStorage(data));
+    }
   }
 
   addToWishlist(id: number) {
@@ -77,7 +86,13 @@ export class CartComponent {
       responseObject.qty = 1;
     }
 
-    this.store.dispatch(new AddToCart(responseObject));
+    if (userId) {
+      this.store.dispatch(new AddToCart(responseObject));
+    } else {
+      product.qty = responseObject.qty;
+      this.store.dispatch(new AddToCartLocalStorage(product));
+    }
+
   };
 
   subractFromItemCount(product: any) {
@@ -93,16 +108,26 @@ export class CartComponent {
 
     const itemFound = this.cartItems.find((item: any) => item.product_id === product.product_id);
     const formattedQty = convertStringToNumber(itemFound.qty);
-    if ((formattedQty-1) === 0) {
+    if ((formattedQty - 1) === 0) {
+      product.qty = 0;
       const object = {
         id: itemFound.id
       };
 
-      this.store.dispatch(new DeleteCart(object));
+      if (userId) {
+        this.store.dispatch(new DeleteCart(object));
+      } else {
+        this.store.dispatch(new SubractFromCartLocalStorage(product));
+      }
     } else {
       requestObject.qty = formattedQty - 1;
 
-      this.store.dispatch(new AddToCart(requestObject));
+      if (userId) {
+        this.store.dispatch(new AddToCart(requestObject));
+      } else {
+        product.qty = requestObject.qty;
+        this.store.dispatch(new SubractFromCartLocalStorage(product));
+      }
     }
   };
