+++ Fastkart-Front-SSR/src/app/shared/components/header/widgets/cart/cart.component.ts
@@ -3,7 +3,7 @@ import { Store, Select } from '@ngxs/store';
 import { Observable } from 'rxjs';
 import { Cart, CartAddOrUpdate } from '../../../../interface/cart.interface';
 import { CartState } from '../../../../state/cart.state';
-import { AddToCart, DeleteCart, GetCartItems, ToggleSidebarCart, UpdateCart } from '../../../../action/cart.action';
+import { AddToCart, AddToCartLocalStorage, DeleteCart, GetCartItems, SubractFromCartLocalStorage, ToggleSidebarCart, UpdateCart } from '../../../../action/cart.action';
 import { ThemeOptionState } from '../../../../state/theme-option.state';
 import { Option } from '../../../../interface/theme-option.interface';
 import { SettingState } from '../../../../state/setting.state';
@@ -21,8 +21,8 @@ import { environment } from 'src/environments/environment';
   styleUrls: ['./cart.component.scss'],
 })
 export class CartComponent {
-  cartItems:Array<any>=[];
-  apiBaseUrl:String=environment.apiBaseUrl;
+  cartItems: Array<any> = [];
+  apiBaseUrl: String = environment.apiBaseUrl;
   @Select(CartState.cartItems) cartItem$: Observable<Cart[]>;
   @Select(CartState.cartTotal) cartTotal$: Observable<number>;
   @Select(CartState.sidebarCartOpen) sidebarCartOpen$: Observable<boolean>;
@@ -56,9 +56,9 @@ export class CartComponent {
       this.setting$.subscribe(setting => this.shippingFreeAmt = setting?.general?.min_order_free_shipping);
       this.cartTotal = total;
       this.shippingCal = (this.cartTotal * 100) / this.shippingFreeAmt;
-      if(this.shippingCal > 100) {
+      if (this.shippingCal > 100) {
         this.shippingCal = 100;
-        if(this.confetti == 0) {
+        if (this.confetti == 0) {
           this.confetti = 1;
           setTimeout(() => {
             this.confetti = 2;
@@ -86,56 +86,56 @@ export class CartComponent {
       variation_id: item?.variation_id ? item?.variation_id : null,
       variation: item?.variation ? item?.variation : null,
       quantity: qty,
-      GuId:null
+      GuId: null
     }
-          // LOGIN NEW CODE 
-          if (userEmailId) {
-            try {
-              const params: CartAddOrUpdate = {
-                id: item?.id,
-                product_id: item?.product?.id,
-                product: item?.product ? item?.product : null,
-                variation_id: item?.variation_id ? item?.variation_id : null,
-                variation: item?.variation ? item?.variation : null,
-                quantity: qty,
-                GuId: null
-              }
-
-
-              this.cartService.addToCart(params).subscribe({
-                next: (response: any) => {  
-                  this.responseError=null;
-                    
-                    localStorage.setItem('AuthToken', response?.token);
-                    localStorage.setItem('UserEmail',response?.Email)
-
-
-                    const redirectUrl =  '/account/dashboard';
-                    this.router.navigateByUrl(redirectUrl);
-                    // window.location.reload();
-                   
-                },
-                error: (error) => {
-                  this.responseError=  error?.error?.messages?.error ?? 'INVAILD EMAILID OR PASSWORD';
-                  console.log("Api Error",error.error.messages.error);
-                },  
-              });
-          } catch (Error: any) {
-            console.log("catch Error",Error);
-          }
-        } else {
-          // Form is invalid, display errors if needed
-          console.log('Invalid form submission');
+    // LOGIN NEW CODE 
+    if (userEmailId) {
+      try {
+        const params: CartAddOrUpdate = {
+          id: item?.id,
+          product_id: item?.product?.id,
+          product: item?.product ? item?.product : null,
+          variation_id: item?.variation_id ? item?.variation_id : null,
+          variation: item?.variation ? item?.variation : null,
+          quantity: qty,
+          GuId: null
         }
-          // END OF LOGIN CODE 
-          // Navigate to the intended URL after successful login
-          // AuthLogin
-          // const redirectUrl = this.authService.redirectUrl || '/account/dashboard';
-        
-          // this.router.navigateByUrl(redirectUrl);
 
-          // Clear the stored redirect URL
-          // this.authService.redirectUrl = undefined; 
+
+        this.cartService.addToCart(params).subscribe({
+          next: (response: any) => {
+            this.responseError = null;
+
+            localStorage.setItem('AuthToken', response?.token);
+            localStorage.setItem('UserEmail', response?.Email)
+
+
+            const redirectUrl = '/account/dashboard';
+            this.router.navigateByUrl(redirectUrl);
+            // window.location.reload();
+
+          },
+          error: (error) => {
+            this.responseError = error?.error?.messages?.error ?? 'INVAILD EMAILID OR PASSWORD';
+            console.log("Api Error", error.error.messages.error);
+          },
+        });
+      } catch (Error: any) {
+        console.log("catch Error", Error);
+      }
+    } else {
+      // Form is invalid, display errors if needed
+      console.log('Invalid form submission');
+    }
+    // END OF LOGIN CODE 
+    // Navigate to the intended URL after successful login
+    // AuthLogin
+    // const redirectUrl = this.authService.redirectUrl || '/account/dashboard';
+
+    // this.router.navigateByUrl(redirectUrl);
+
+    // Clear the stored redirect URL
+    // this.authService.redirectUrl = undefined; 
   }
 
 
@@ -153,7 +153,16 @@ export class CartComponent {
   // }
 
   delete(id: number) {
-    this.store.dispatch(new DeleteCart(id));
+    const user_id = getStringDataFromLocalStorage("user_id");
+
+    if (user_id) {
+      this.store.dispatch(new DeleteCart({ id: id }));
+    } else {
+      let data = this.cartItems.find((e: any) => e.id === id);
+      data.qty = 0;
+
+      this.store.dispatch(new SubractFromCartLocalStorage(data));
+    }
   }
 
   addCart(product: any) {
@@ -175,7 +184,12 @@ export class CartComponent {
       responseObject.qty = 1;
     }
 
-    this.store.dispatch(new AddToCart(responseObject));
+    if (userId) {
+      this.store.dispatch(new AddToCart(responseObject));
+    } else {
+      product.qty = responseObject.qty;
+      this.store.dispatch(new AddToCartLocalStorage(product));
+    }
   };
 
   subractItemCount(product: any) {
@@ -190,16 +204,26 @@ export class CartComponent {
     };
     const itemFound = this.cartItems.find((item: any) => item.product_id === product.product_id);
     const formattedQty = convertStringToNumber(itemFound.qty);
-    if ((formattedQty-1) === 0) {
+    if ((formattedQty - 1) === 0) {
+      product.qty = 0;
       const object = {
         id: itemFound.id
       };
 
-      this.store.dispatch(new DeleteCart(object));
+      if (userId) {
+        this.store.dispatch(new DeleteCart(object));
+      } else {
+        this.store.dispatch(new SubractFromCartLocalStorage(product));
+      }
     } else {
       requestObject.qty = formattedQty - 1;
 
-      this.store.dispatch(new AddToCart(requestObject));
+      if (userId) {
+        this.store.dispatch(new AddToCart(requestObject));
+      } else {
+        product.qty = requestObject.qty;
+        this.store.dispatch(new SubractFromCartLocalStorage(product));
+      }
     }
   };
